Initial start to creating slice reducers.
